backend: fix the decoder v4l2 buffer assignment
I thought the decoder would never be used, but when I test with the
1080P video, I find a va surface which is not the render target
been processed would be send to vaDeriveImage(). I didn't meet this
with 720P video yet.

The simple way to solve this problem is just return an error status,
then the VA-API would fallback to vaGetImage(). I didn't see much
performance effect here.

Signed-off-by: Randy Li <randy.li@rock-chips.com>